File structure changed
Core file has now been separated in different files on the lib folder
